warnings ...

+++ scilab/modules/scicos/src/c/ezxml.c
@@ -632,7 +632,7 @@ ezxml_t ezxml_parse_fp(FILE *fp)
 
     if (! s) return NULL;
     root = (ezxml_root_t)ezxml_parse_str(s, len);
-    root->len = -1; // so we know to free s in ezxml_free()
+    root->len = (size_t)-1; // so we know to free s in ezxml_free()
     return &root->xml;
 }
 
@@ -660,7 +660,7 @@ ezxml_t ezxml_parse_fd(int fd)
 #endif // EZXML_NOMMAP
         l = read(fd, m = malloc(st.st_size), st.st_size);
         root = (ezxml_root_t)ezxml_parse_str(m, l);
-        root->len = -1; // so we know to free s in ezxml_free()
+        root->len = (size_t)-1; // so we know to free s in ezxml_free()
 #ifndef EZXML_NOMMAP
     }
 #endif // EZXML_NOMMAP
@@ -725,7 +725,7 @@ char *ezxml_toxml_r(ezxml_t xml, char **s, size_t *len, size_t *max,
             *s = realloc(*s, *max += EZXML_BUFSIZE);
 
         *len += sprintf(*s + *len, " %s=\"", xml->attr[i]);
-        ezxml_ampencode(xml->attr[i + 1], -1, s, len, max, 1);
+        ezxml_ampencode(xml->attr[i + 1], (size_t)(-1), s, len, max, 1);
         *len += sprintf(*s + *len, "\"");
     }
 
@@ -737,13 +737,13 @@ char *ezxml_toxml_r(ezxml_t xml, char **s, size_t *len, size_t *max,
             *s = realloc(*s, *max += EZXML_BUFSIZE);
 
         *len += sprintf(*s + *len, " %s=\"", attr[i][j]);
-        ezxml_ampencode(attr[i][j + 1], -1, s, len, max, 1);
+        ezxml_ampencode(attr[i][j + 1], (size_t)(-1), s, len, max, 1);
         *len += sprintf(*s + *len, "\"");
     }
     *len += sprintf(*s + *len, ">");
 
     *s = (xml->child) ? ezxml_toxml_r(xml->child, s, len, max, 0, attr) //child
-                      : ezxml_ampencode(xml->txt, -1, s, len, max, 0);  //data
+                      : ezxml_ampencode(xml->txt, (size_t)(-1), s, len, max, 0);  //data
     
     while (*len + strlen(xml->name) + 4 > *max) // reallocate s
         *s = realloc(*s, *max += EZXML_BUFSIZE);
@@ -752,7 +752,7 @@ char *ezxml_toxml_r(ezxml_t xml, char **s, size_t *len, size_t *max,
 
     while (txt[off] && off < xml->off) off++; // make sure off is within bounds
     return (xml->ordered) ? ezxml_toxml_r(xml->ordered, s, len, max, off, attr)
-                          : ezxml_ampencode(txt + off, -1, s, len, max, 0);
+                          : ezxml_ampencode(txt + off, (size_t)(-1), s, len, max, 0);
 }
 
 // Converts an ezxml structure back to xml. Returns a string of xml data that
